making progress at moving the legislator profile into the header
