+++ app/src/main/java/com/sunlightlabs/android/congress/LegislatorLoader.java
@@ -24,7 +24,7 @@ public class LegislatorLoader extends Activity implements LoadsLegislator {
 		
 		Intent i = getIntent();
 		id = i.getStringExtra("id");
-		intent = (Intent) i.getParcelableExtra("intent");
+		intent = i.getParcelableExtra("intent");
 		
 		// if coming from a shortcut intent, there appears to be a bug with packaging sub-intents
 		// and the intent will be null
