+++ app/src/main/java/com/sunlightlabs/android/congress/LegislatorPager.java
@@ -1,8 +1,13 @@
 package com.sunlightlabs.android.congress;
 
+import android.Manifest;
 import android.app.Activity;
+import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.content.res.Resources;
 import android.database.Cursor;
+import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.ContactsContract;
@@ -10,25 +15,30 @@ import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.Button;
+import android.widget.ImageView;
+import android.widget.TextView;
 
 import com.sunlightlabs.android.congress.fragments.BillListFragment;
+import com.sunlightlabs.android.congress.fragments.CommitteeListFragment;
 import com.sunlightlabs.android.congress.fragments.LegislatorLoaderFragment;
-import com.sunlightlabs.android.congress.fragments.LegislatorProfileFragment;
 import com.sunlightlabs.android.congress.fragments.RollListFragment;
+import com.sunlightlabs.android.congress.tasks.LoadPhotoTask;
 import com.sunlightlabs.android.congress.utils.ActionBarUtils;
 import com.sunlightlabs.android.congress.utils.ActionBarUtils.HasActionMenu;
 import com.sunlightlabs.android.congress.utils.Analytics;
 import com.sunlightlabs.android.congress.utils.Database;
+import com.sunlightlabs.android.congress.utils.LegislatorImage;
 import com.sunlightlabs.android.congress.utils.TitlePageAdapter;
 import com.sunlightlabs.android.congress.utils.Utils;
 import com.sunlightlabs.congress.models.CongressException;
 import com.sunlightlabs.congress.models.Legislator;
 
-public class LegislatorPager extends Activity implements HasActionMenu {
+public class LegislatorPager extends Activity implements HasActionMenu, LoadPhotoTask.LoadsPhoto {
 	public String bioguide_id;
 	public Legislator legislator;
 	public String tab;
+
+    private Drawable avatar;
 	
 	public Database database;
 	public Cursor cursor;
@@ -37,8 +47,10 @@ public class LegislatorPager extends Activity implements HasActionMenu {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-		setContentView(R.layout.pager_titled);
+		setContentView(R.layout.pager_legislator);
+
         Analytics.init(this);
+        Utils.setupAPI(this);
 		
 		Bundle extras = getIntent().getExtras();
 		bioguide_id = extras.getString("bioguide_id");
@@ -46,10 +58,13 @@ public class LegislatorPager extends Activity implements HasActionMenu {
 		tab = extras.getString("tab");
 		
 		setupControls();
-		
+
+        // currently unused - we always use LegislatorLoaderFragment to get in here
 		if (legislator == null)
 			LegislatorLoaderFragment.start(this);
-		else
+
+        // in practice, always the route
+        else
 			onLoadLegislator(legislator);
 	}
 	
@@ -61,7 +76,13 @@ public class LegislatorPager extends Activity implements HasActionMenu {
 		
 		setupDatabase();
 		setupButtons();
-		setupPager();	
+        setupProfile();
+		setupPager();
+
+        if (avatar != null)
+            displayAvatar();
+        else
+            loadPhoto();
 	}
 	
 	public void onLoadLegislator(CongressException exception) {
@@ -82,19 +103,18 @@ public class LegislatorPager extends Activity implements HasActionMenu {
 		findViewById(R.id.pager_container).setVisibility(View.GONE);
 		Utils.setLoading(this, R.string.legislator_loading);
 		
-		((Button) findViewById(R.id.refresh)).setOnClickListener(new View.OnClickListener() {
+		findViewById(R.id.refresh).setOnClickListener(new View.OnClickListener() {
 			public void onClick(View v) {
 				refresh();
 			}
 		});
-	}
+    }
 	
 	private void setupPager() {
 		TitlePageAdapter adapter = new TitlePageAdapter(this);
-		adapter.add("info", R.string.tab_profile, LegislatorProfileFragment.create(legislator));
 		adapter.add("votes", R.string.tab_votes, RollListFragment.forLegislator(legislator));
 		adapter.add("bills", R.string.tab_bills, BillListFragment.forSponsor(legislator));
-		
+		adapter.add("committees", R.string.committees, CommitteeListFragment.forLegislator(legislator));
 		if (tab != null) adapter.selectPage(tab);
 	}
 	
@@ -128,6 +148,115 @@ public class LegislatorPager extends Activity implements HasActionMenu {
 		
 		ActionBarUtils.setActionMenu(this, R.menu.legislator);
 	}
+
+	public void setupProfile() {
+        if (!legislator.in_office)
+            findViewById(R.id.out_of_office_text).setVisibility(View.VISIBLE);
+
+        String party = partyName(legislator.party);
+        String state = Utils.stateCodeToName(this, legislator.state);
+        ((TextView) findViewById(R.id.profile_state_party)).setText(party + " from " + state);
+
+        String domain = legislator.getDomain();
+        if (legislator.leadership_role != null && !legislator.leadership_role.equals(""))
+            domain = legislator.leadership_role + ", " + domain;
+        ((TextView) findViewById(R.id.profile_domain)).setText(domain);
+
+        socialButton(R.id.twitter, legislator.twitterUrl(), Analytics.LEGISLATOR_TWITTER);
+        socialButton(R.id.youtube, legislator.youtubeUrl(), Analytics.LEGISLATOR_YOUTUBE);
+        socialButton(R.id.facebook, legislator.facebookUrl(), Analytics.LEGISLATOR_FACEBOOK);
+
+        TextView officeView = (TextView) findViewById(R.id.profile_office);
+        if (legislator.office != null && !legislator.office.equals(""))
+            officeView.setText(officeName(legislator.office));
+        else
+            officeView.setVisibility(View.GONE);
+
+        // allow for devices without phones
+        boolean hasPhone = getPackageManager().hasSystemFeature(PackageManager.FEATURE_TELEPHONY);
+        if (hasPhone && legislator.phone != null && !legislator.phone.equals("")) {
+            findViewById(R.id.call_office).setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    callOffice();
+                }
+            });
+        } else
+            findViewById(R.id.call_office_container).setVisibility(View.GONE);
+
+        if (legislator.website != null && !legislator.website.equals("")) {
+            findViewById(R.id.visit_website).setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    visit(legislator.website, Analytics.LEGISLATOR_WEBSITE);
+                }
+            });
+        } else
+            findViewById(R.id.visit_website_container).setVisibility(View.GONE);
+    }
+
+    public void loadPhoto() {
+        new LoadPhotoTask(this, LegislatorImage.PIC_LARGE).execute(legislator.bioguide_id);
+    }
+
+    @Override
+    public void onLoadPhoto(Drawable avatar, Object tag) {
+        if (avatar == null) {
+            Resources resources = getResources();
+            if (resources != null)
+                avatar = resources.getDrawable(R.drawable.person);
+        }
+        this.avatar = avatar;
+        displayAvatar();
+    }
+
+    public void displayAvatar() {
+        ((ImageView) findViewById(R.id.profile_picture)).setImageDrawable(avatar);
+    }
+
+    public void callOffice() {
+        Analytics.legislatorCall(this, legislator.bioguide_id);
+
+        // if user gave us permission to directly initiate calls, do so
+        if (Utils.checkPermission(this, Manifest.permission.CALL_PHONE))
+            startActivity(new Intent(Intent.ACTION_CALL, Uri.parse("tel://" + legislator.phone)));
+        // otherwise, open up the dialer with the number ready to go (needs no permission)
+        else
+            startActivity(new Intent(Intent.ACTION_DIAL));
+    }
+
+    public void visit(String url, String social) {
+        Analytics.legislatorWebsite(this, legislator.bioguide_id, social);
+        startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(url)));
+    }
+
+    private void socialButton(int id, final String url, final String network) {
+        View view = findViewById(id);
+        if (url != null && !url.equals("")) {
+            view.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View arg0) {
+                    visit(url, network);
+                }
+            });
+        } else
+            view.setVisibility(View.GONE);
+    }
+
+    public static String partyName(String code) {
+        if (code.equals("D"))
+            return "Democrat";
+        if (code.equals("R"))
+            return "Republican";
+        if (code.equals("I"))
+            return "Independent";
+        else
+            return "";
+    }
+
+    public static String officeName(String office) {
+        return office.replaceAll("(?:House|Senate)? ?(?:Office)? ?Building", "").trim();
+    }
 
 	private void toggleFavoriteStar(boolean enabled) {
 		if (enabled)
@@ -207,4 +336,9 @@ public class LegislatorPager extends Activity implements HasActionMenu {
 		super.onStop();
 		Analytics.stop(this);
 	}
+
+	@Override
+    public Context getContext() {
+        return this;
+    }
 }
