+++ app/src/main/res/layout/pager_legislator.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    >
+    
+    <include layout="@layout/legislator_profile" />
+    
+    <include layout="@layout/pager" />
+    
+    <include layout="@layout/list_empty"
+		android:layout_width="match_parent"
+	    android:layout_height="match_parent"
+	    android:visibility="gone"
+	    />
+	
+</LinearLayout>
